requirements.txt
Adafruit-Blinka==8.18.0
adafruit-circuitpython-busdevice==5.2.4
adafruit-circuitpython-neopixel==6.3.8
adafruit-circuitpython-pixelbuf==2.0.0
adafruit-circuitpython-requests==1.13.1
adafruit-circuitpython-typing==1.9.0
Adafruit-PlatformDetect==3.45.0
Adafruit-PureIO==1.1.10
black==23.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
Flask==2.2.3
gevent==22.10.2
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
platformdirs==3.1.0
pyftdi==0.54.0
pyrf24==0.2.2
pyserial==3.5
pyusb==1.2.1
requests==2.28.2
rpi-ws281x==4.3.4
RPi.GPIO==0.7.1
sysv-ipc==1.1.0
tomli==2.0.1
typing-extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
zipp==3.15.0
zope.event==4.6
zope.interface==6.0
